pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=1.54.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
